feat: ERNIE LoRA training — fix weight keys, gradient checkpointing, RoPE cache
Key fixes:
- Fix LoRA save/load key prefix: diffusion_model.* → transformer.*
- Split ErnieLatentCreator.pack_latents + bn_normalize_latents (were fused)
- AdaLN modulation: remove unnecessary float32 casts in transformer_block
- mx.eval placement fix in gradient accumulation (only after full accumulation)
- mx.clear_cache in low_ram mode only when not mid-accumulation
- ERNIE initializer: eval+clear_cache after weight load, before LoRA apply
- requires_sigma_shift=True for ernie-image/ernie-image-turbo
- lora_training_steps/lora_training_guidance on ModelConfig for canonical per-model values
- Runner warns when user config steps/guidance are overridden by canonical model values

New features:
- Gradient checkpointing: mx.checkpoint per transformer block (gradient_checkpointing, gradient_checkpointing_every_n)
- RoPE + attention mask cache in ErnieTransformer (_pos_cache, keyed by H/W/T/text_lens)
- cache_limit_gb: cap MLX free-buffer pool to prevent swap spikes on large models
- skip_initial_preview: skip preview image generation at step 0
- LoRA mapping: add time_embedding.linear_1/2 as ERNIE targets; add transformer.* prefix patterns

+++ src/mflux/models/ernie_image/model/ernie_transformer/transformer_block.py
@@ -1,5 +1,5 @@
-import mlx.core as mx
 from mlx import nn
+import mlx.core as mx
 
 from mflux.models.ernie_image.model.ernie_transformer.attention import ErnieAttention
 from mflux.models.ernie_image.model.ernie_transformer.feed_forward import ErnieFeedForward
@@ -25,13 +25,13 @@ class ErnieTransformerBlock(nn.Module):
         # Self-attention with AdaLN pre-norm
         residual = x
         x = self.adaLN_sa_ln(x)
-        x = (x.astype(mx.float32) * (1 + scale_msa.astype(mx.float32)) + shift_msa.astype(mx.float32)).astype(residual.dtype)
-        x = residual + (gate_msa.astype(mx.float32) * self.self_attention(x, freqs_cis, mask).astype(mx.float32)).astype(residual.dtype)
+        x = x * (1 + scale_msa) + shift_msa
+        x = residual + gate_msa * self.self_attention(x, freqs_cis, mask)
 
         # FFN with AdaLN pre-norm
         residual = x
         x = self.adaLN_mlp_ln(x)
-        x = (x.astype(mx.float32) * (1 + scale_mlp.astype(mx.float32)) + shift_mlp.astype(mx.float32)).astype(residual.dtype)
-        x = residual + (gate_mlp.astype(mx.float32) * self.mlp(x).astype(mx.float32)).astype(residual.dtype)
+        x = x * (1 + scale_mlp) + shift_mlp
+        x = residual + gate_mlp * self.mlp(x)
 
         return x
